CREATE: 회원가입, 게시글 업데이트, 삭제수정

+++ spartamarket/products/views.py
@@ -1,8 +1,10 @@
+from django.http import Http404
 from django.shortcuts import get_object_or_404, redirect, render
-
 from products.form import ProductForm
 from products.models import Product
-from django.views.decorators.http import require_POST
+from django.views.decorators.http import require_POST, require_http_methods
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 
 def index(request):
@@ -10,18 +12,20 @@ def index(request):
     return render(request, 'products/index.html', {"products": products})
 
 
+@login_required
 def list_view(request):
     products = Product.objects.all()
     print(products)
     return render(request, 'products/list.html', {"products": products})
 
 
+@login_required
 def create_view(request):
     form = ProductForm()
     if request.method == 'POST':
         form = ProductForm(request.POST)
         if form.is_valid():
-            form.save()
+            form.save(author=request.user)
             return redirect("products:list")
     return render(request, 'products/create.html', {"form": form})
 
@@ -31,19 +35,35 @@ def detail_view(request, product_id):
     return render(request, 'products/detail.html', {"product": product})
 
 
+@login_required
+@require_http_methods(["GET", "POST"])
 def update_view(request, product_id):
     product = get_object_or_404(Product, pk=product_id)
+    if product.author != request.user:
+        messages.warning(request, '작성자 본인만 수정이 가능합니다.')
+        return redirect("products:detail", product_id=product_id)
     form = ProductForm(request.POST or None, instance=product)
     if request.method == 'POST':
-        form = ProductForm(request.POST)
+        form = ProductForm(request.POST, instance=product)
         if form.is_valid():
             form.save()
             return redirect("products:list")
     return render(request, 'products/create.html', {"form": form})
 
 
+@login_required
 @require_POST
 def delete_view(request, product_id):
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
+    if product.author == request.user:
+        product.delete()
+    return redirect("products:list")
+
+
+@login_required
+@require_POST
+def good_view(request, product_id):
+    product = get_object_or_404(Product, id=product_id)
+    user = request.user
+    user.good_products.add(product)
     return redirect("products:list")
